moving to theme directory
