+++ app/src/main/java/org/mozilla/social/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import org.mozilla.social.core.designsystem.MozillaSocialTheme
+import org.mozilla.social.core.designsystem.theme.MozillaSocialTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
