+++ core/designsystem/src/main/java/org/mozilla/social/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package org.mozilla.social.core.designsystem
+package org.mozilla.social.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
